+++ src/Component/TextField/index.tsx
@@ -1,7 +1,7 @@
-import React, {useMemo} from 'react';
-import {IconButton, IconProps, InputAdornment, TextField as MuiTextField, TextFieldProps} from '@mui/material';
+import React, { useMemo } from 'react';
+import { IconButton, IconProps, InputAdornment, TextField as MuiTextField, TextFieldProps } from '@mui/material';
 import Clear from '@mui/icons-material/Clear';
-import {IconButtonProps} from '@mui/material/IconButton';
+import { IconButtonProps } from '@mui/material/IconButton';
 
 export type Props = TextFieldProps & {
   clearButtonSize?: IconButtonProps['size'];
@@ -9,33 +9,50 @@ export type Props = TextFieldProps & {
   clearIconSize?: IconButtonProps['size'] & IconProps['fontSize']
   hideClearButton?: boolean;
   forceEnableClearButton?: boolean;
+  selectOnFocus?: boolean;
 };
 const TextField = React.forwardRef<HTMLDivElement, Props>(({
-                                                             clearButtonSize = 'small',
-                                                             clearIconSize = 'small',
-                                                             value = null,
-                                                             hideClearButton,
-                                                             forceEnableClearButton,
-                                                             clearButtonProps,
-                                                             name,
-                                                             ...props
-                                                           }, ref) => {
+  clearButtonSize = 'small',
+  clearIconSize = 'small',
+  value = null,
+  hideClearButton,
+  forceEnableClearButton,
+  clearButtonProps,
+  name,
+  selectOnFocus,
+  onFocus,
+  ...props
+}, ref) => {
   const isDirty = value !== null && value !== undefined && value !== '';
   const clear = React.useCallback(
     () => {
-      props?.onChange?.({target: {value: null}} as React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>)
+      props?.onChange?.({ target: { value: null } } as React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>)
     },
     [props.onChange]
   )
+  const _onFocus: TextFieldProps['onFocus'] = React.useCallback((e) => {
+    if (onFocus) {
+      onFocus(e)
+    }
+    if (selectOnFocus) {
+      const target = e.target;
+      window.requestAnimationFrame(() => {
+        if (target) {
+          target.select();
+        }
+      });
+    }
+  }, [selectOnFocus, onFocus])
   const endAdornmentClearButton = React.useMemo(
     () => (forceEnableClearButton || isDirty) && !hideClearButton ? (
-        <IconButton
-          disabled={props.disabled && !forceEnableClearButton}
-          size={clearButtonSize}
-          color={'secondary'}
-          onClick={clear}
-          {...clearButtonProps}
-        ><Clear fontSize={clearIconSize}/></IconButton>)
+      <IconButton
+        disabled={props.disabled && !forceEnableClearButton}
+        size={clearButtonSize}
+        color={'secondary'}
+        onClick={clear}
+        tabIndex={-1}
+        {...clearButtonProps}
+      ><Clear fontSize={clearIconSize} /></IconButton>)
       : <></>,
     [isDirty, props.disabled, clearButtonProps, clearButtonSize, forceEnableClearButton, clearIconSize, clear, hideClearButton]
   )
@@ -52,7 +69,7 @@ const TextField = React.forwardRef<HTMLDivElement, Props>(({
   );
   const onChange: TextFieldProps['onChange'] = React.useCallback(
     (e) => {
-      props.onChange(e)
+      props?.onChange?.(e)
     },
     [props.onChange]
   )
@@ -61,6 +78,7 @@ const TextField = React.forwardRef<HTMLDivElement, Props>(({
     name={name}
     value={value || ''}
     {...props}
+    onFocus={_onFocus}
     onChange={onChange}
     InputProps={preparedInputProps}
   />