+++ src/Showcase/Component/Forms.showcase.tsx
@@ -1,48 +1,59 @@
+/* eslint-disable no-console */
 /* eslint-disable max-lines */
 import React from 'react';
 import AppNavigation from './AppNavigation.showcase';
 import {
   Box,
+  Button,
   Card,
   CardContent,
   CardHeader,
   Checkbox,
   Container,
+  Divider,
   FormControl,
   FormControlLabel,
-  FormGroup, FormHelperText,
+  FormGroup,
+  FormHelperText,
   FormLabel,
-  Grid, InputLabel, MenuItem,
+  Grid,
+  InputAdornment,
+  InputLabel,
+  MenuItem,
   Radio,
   RadioGroup,
+  Switch,
+  TextField,
   Typography,
-  Switch, Divider, TextField, InputAdornment
 } from '@mui/material';
 import CodeBox from './CodeBox.showcase';
 import WarningIcon from '@mui/icons-material/Warning';
 import PriceField from '../../Component/PriceInput';
-import {DatePicker, TimePicker} from '@mui/x-date-pickers';
+import { DatePicker, TimePicker } from '@mui/x-date-pickers';
 import 'date-fns';
 import KeyboardDatePicker from '../../Component/DatePicker';
-import Select, {type Props as SelectProps} from '../../Component/Select';
+import Select, { type Props as SelectProps } from '../../Component/Select';
 import TextFieldClearButton from '../../Component/TextField';
 import definition from '../../Component/App/Stateless/Style/Theme/Definition';
 import Layout from '../../Component/App/Layout';
-import {DatePickerProps} from '@mui/lab';
+import { DatePickerProps } from '@mui/lab';
+import { LANGUAGE } from '../../Types/Language';
 
 const Forms = () => {
-  const [checkboxState, setCheckboxState] = React.useState({checkboxA: true, checkboxB: false, checkboxC: false, checkboxD: false});
-  const [switchState, setSwitchState] = React.useState({switchA: true, switchB: false, switchC: false, switchD: false});
+  const [checkboxState, setCheckboxState] = React.useState({ checkboxA: true, checkboxB: false, checkboxC: false, checkboxD: false });
+  const textFieldDEId = 'textFieldDEId';
+  const textFieldENId = 'textFieldENId';
+  const [switchState, setSwitchState] = React.useState({ switchA: true, switchB: false, switchC: false, switchD: false });
   const [radioValue, setRadioValue] = React.useState('radio1');
   const [displayRowStyle, setDisplayRowStyle] = React.useState(false);
   const [showError, setShowError] = React.useState(false);
   const [selectValue, setSelectValue] = React.useState<string>('');
-  const [priceValue, setPriceValue] = React.useState('1000.55');
+  const [priceValue, setPriceValue] = React.useState<number | undefined>(123466);
   const [selectedDate, setSelectedDate] = React.useState<Date | null>(new Date('2021-07-09T10:00:00'));
   const [textFieldValue, setTextFieldValue] = React.useState('This is a default value');
   const [dateRange, setDateRange] = React.useState<[Date, Date]>([new Date('2021-07-09T10:00:00'), new Date('2021-07-09T10:00:00')]);
-  const handleCheckboxChange = (event: React.ChangeEvent<HTMLInputElement>) => setCheckboxState({...checkboxState, [event.target.name]: event.target.checked});
-  const handleSwitchChange = (name: string, value: boolean) => setSwitchState({...switchState, [name]: value});
+  const handleCheckboxChange = (event: React.ChangeEvent<HTMLInputElement>) => setCheckboxState({ ...checkboxState, [event.target.name]: event.target.checked });
+  const handleSwitchChange = (name: string, value: boolean) => setSwitchState({ ...switchState, [name]: value });
   const handleRadioChange = (event: React.ChangeEvent<HTMLInputElement>) => setRadioValue((event.target as HTMLInputElement).value);
   const toggleDisplayRowStyle = () => setDisplayRowStyle(!displayRowStyle);
   const toggleShowError = () => setShowError(!showError);
@@ -50,17 +61,17 @@ const Forms = () => {
   const handleDateChange: DatePickerProps<Date>['onChange'] = (date: Date) => setSelectedDate(date);
 
   return (
-    <Layout navigation={<AppNavigation/>}>
+    <Layout navigation={<AppNavigation />}>
       {/* <MuiPickersUtilsProvider utils={DateFnsUtils}> */}
       <Container disableGutters maxWidth={false}>
         <Card>
           <CardHeader title='Forms'
-                      subheader={<>Errors are controlled by the <CodeBox component={'span'} size={'small'}>{`<FormControl required error>...</FormControl>`}</CodeBox> Element.
-                        The <CodeBox component={'span'} size={'small'}>required</CodeBox> attribute controls the asterisk (*) char.</>}/>
+            subheader={<>Errors are controlled by the <CodeBox component={'span'} size={'small'}>{`<FormControl required error>...</FormControl>`}</CodeBox> Element.
+              The <CodeBox component={'span'} size={'small'}>required</CodeBox> attribute controls the asterisk (*) char.</>} />
           <CardContent>
             <Grid container spacing={1}>
-              <Grid item><FormControlLabel control={<Switch onChange={toggleDisplayRowStyle} checked={displayRowStyle}/>} label='Enable row style'/></Grid>
-              <Grid item><FormControlLabel control={<Switch onChange={toggleShowError} checked={showError}/>} label='Show error'/></Grid>
+              <Grid item><FormControlLabel control={<Switch onChange={toggleDisplayRowStyle} checked={displayRowStyle} />} label='Enable row style' /></Grid>
+              <Grid item><FormControlLabel control={<Switch onChange={toggleShowError} checked={showError} />} label='Show error' /></Grid>
             </Grid>
             <Box mt={2}>
               <Grid container spacing={displayRowStyle ? 4 : 2}>
@@ -69,20 +80,20 @@ const Forms = () => {
                     <FormLabel>Checkbox Example</FormLabel>
                     <FormGroup row={displayRowStyle}>
                       <FormControlLabel
-                        control={<Checkbox checked={checkboxState.checkboxA} onChange={handleCheckboxChange} name='checkboxA'/>}
+                        control={<Checkbox checked={checkboxState.checkboxA} onChange={handleCheckboxChange} name='checkboxA' />}
                         label='Checkbox A'
                       />
                       <FormControlLabel
-                        control={<Checkbox checked={checkboxState.checkboxB} onChange={handleCheckboxChange} name='checkboxB'/>}
+                        control={<Checkbox checked={checkboxState.checkboxB} onChange={handleCheckboxChange} name='checkboxB' />}
                         label='Checkbox B'
                       />
                       <FormControlLabel
-                        control={<Checkbox checked={checkboxState.checkboxC} onChange={handleCheckboxChange} name='checkboxC'/>}
+                        control={<Checkbox checked={checkboxState.checkboxC} onChange={handleCheckboxChange} name='checkboxC' />}
                         label='Checkbox C'
                       />
                       <FormControlLabel
                         disabled
-                        control={<Checkbox checked={checkboxState.checkboxD} onChange={handleCheckboxChange} name='checkboxD'/>}
+                        control={<Checkbox checked={checkboxState.checkboxD} onChange={handleCheckboxChange} name='checkboxD' />}
                         label='Checkbox D'
                       />
                     </FormGroup>
@@ -93,10 +104,10 @@ const Forms = () => {
                   <FormControl required error={showError}>
                     <FormLabel>Radio Buttons Example</FormLabel>
                     <RadioGroup row={displayRowStyle} name='radio' value={radioValue} onChange={handleRadioChange}>
-                      <FormControlLabel value='radio1' control={<Radio/>} label='Radio 1'/>
-                      <FormControlLabel value='radio2' control={<Radio/>} label='Radio 2'/>
-                      <FormControlLabel value='radio3' control={<Radio/>} label='Radio 3'/>
-                      <FormControlLabel value='radio4' disabled control={<Radio/>} label='Radio 4'/>
+                      <FormControlLabel value='radio1' control={<Radio />} label='Radio 1' />
+                      <FormControlLabel value='radio2' control={<Radio />} label='Radio 2' />
+                      <FormControlLabel value='radio3' control={<Radio />} label='Radio 3' />
+                      <FormControlLabel value='radio4' disabled control={<Radio />} label='Radio 4' />
                     </RadioGroup>
                     <FormHelperText>{showError ? 'Ups, an error was detected!' : 'This is a helper text'}</FormHelperText>
                   </FormControl>
@@ -110,31 +121,31 @@ const Forms = () => {
                           checked={switchState.switchA}
                           onChange={(event, value) => handleSwitchChange('switchA', value)} name='switchA'
                         />}
-                        label='Switch 1'/>
+                        label='Switch 1' />
                       <FormControlLabel
                         control={<Switch
                           checked={switchState.switchB}
                           onChange={(event, value) => handleSwitchChange('switchB', value)} name='switchB'
                         />}
-                        label='Switch 2'/>
+                        label='Switch 2' />
                       <FormControlLabel
                         control={<Switch
                           checked={switchState.switchC}
                           onChange={(event, value) => handleSwitchChange('switchC', value)} name='switchC'
                         />}
-                        label='Switch 3'/>
+                        label='Switch 3' />
                       <FormControlLabel
                         control={<Switch
                           checked={switchState.switchD}
                           onChange={(event, value) => handleSwitchChange('switchD', value)} name='switchD'
                         />}
-                        label='Switch 4'/>
+                        label='Switch 4' />
                     </FormGroup>
                     <FormHelperText>{showError ? 'Ups, an error was detected!' : 'This is a helper text'}</FormHelperText>
                   </FormControl>
                 </Grid>
                 <Grid item xs={12}>
-                  <Divider/>
+                  <Divider />
                   <Box mt={2}>
                     <Typography variant={'h6'}>Select/Dropdown Elements</Typography>
                     <Typography color={'textSecondary'}>Icons in <CodeBox component={'span'} size={'small'}>Select</CodeBox>'s need a little custom styling: <CodeBox
@@ -144,7 +155,7 @@ const Forms = () => {
                   <Box mt={2}>
                     <Grid container spacing={1}>
                       <Grid item xs={6}>
-                        <FormControl sx={{width: 200}} color='primary' required error={showError}>
+                        <FormControl sx={{ width: 200 }} color='primary' required error={showError}>
                           <Select<string>
                             value={selectValue}
                             label={'Choose Value'}
@@ -152,18 +163,18 @@ const Forms = () => {
                             canClear
                           >
                             <MenuItem value={'1'}><Box alignItems={'center'} display={'flex'}>
-                              <WarningIcon sx={{fontSize: definition.typography.pxToRem(20)}}/>
+                              <WarningIcon sx={{ fontSize: definition.typography.pxToRem(20) }} />
                               <Box component={'span'} ml={1}>Value 1</Box>
                             </Box></MenuItem>
                             <MenuItem value={'2'}><Box alignItems={'center'} display={'flex'}>
-                              <WarningIcon sx={{fontSize: definition.typography.pxToRem(20)}} color={'primary'}/>
+                              <WarningIcon sx={{ fontSize: definition.typography.pxToRem(20) }} color={'primary'} />
                               <Box component={'span'} ml={1} color={definition.palette.primary.main}>Value 2</Box>
                             </Box></MenuItem>
                             <MenuItem value={'3'} disabled>Value 3</MenuItem>
                             <MenuItem value={'4'}>Value 4 with a longer label</MenuItem>
                             <MenuItem value={'5'}>Value 5 with even a more longer, longer and longer label</MenuItem>
                             <MenuItem value={'6'} disabled><Box alignItems={'center'} display={'flex'}>
-                              <WarningIcon sx={{fontSize: definition.typography.pxToRem(20)}} color={'primary'}/>
+                              <WarningIcon sx={{ fontSize: definition.typography.pxToRem(20) }} color={'primary'} />
                               <Box component={'span'} ml={1} color={definition.palette.primary.main}>Value 6</Box>
                             </Box></MenuItem>
                             <MenuItem value={'7'}>Value 7</MenuItem>
@@ -180,7 +191,7 @@ const Forms = () => {
                         </FormControl>
                       </Grid>
                       <Grid>
-                        <FormControl sx={{minWidth: 200}}>
+                        <FormControl sx={{ minWidth: 200 }}>
                           <InputLabel shrink={selectValue !== ''}>Disabled Element</InputLabel>
                           {/*eslint-disable-next-line no-console */}
                           <Select<false> value={''} onChange={console.log} disabled>
@@ -210,8 +221,8 @@ const Forms = () => {
                           defaultValue='0.5'
                           helperText='This is a carat text field which has an endAdornment and the additional attributes: min, max and step'
                           type={'number'}
-                          InputProps={{endAdornment: <InputAdornment position='end' sx={{userSelect: 'none'}}>ct</InputAdornment>}}
-                          inputProps={{min: 0, max: 1, step: '0.01'}}
+                          InputProps={{ endAdornment: <InputAdornment position='end' sx={{ userSelect: 'none' }}>ct</InputAdornment> }}
+                          inputProps={{ min: 0, max: 1, step: '0.01' }}
                         />
                       </Grid>
                       <Grid item xs={2}>
@@ -221,8 +232,8 @@ const Forms = () => {
                           defaultValue='50'
                           helperText='This is a weight text field which has an endAdornment and the additional attributes: min, max and step'
                           type={'number'}
-                          InputProps={{endAdornment: <InputAdornment position='end' sx={{userSelect: 'none'}}>gr</InputAdornment>}}
-                          inputProps={{min: '0', max: '100', step: '1'}}
+                          InputProps={{ endAdornment: <InputAdornment position='end' sx={{ userSelect: 'none' }}>gr</InputAdornment> }}
+                          inputProps={{ min: '0', max: '100', step: '1' }}
                         />
                       </Grid>
                       <Grid item xs={2}>
@@ -235,23 +246,14 @@ const Forms = () => {
                         />
                       </Grid>
                       <Grid item xs={2}>
-                        <PriceField
-                          error={showError}
-                          label='Core Price field'
-                          value={priceValue}
-                          helperText='This is a price field which is provided by the core'
-                          onChange={event => setPriceValue(event.target.value)}
-                        />
-                      </Grid>
-                      <Grid item xs={2}>
                         <DatePicker
                           slotProps={{
                             textField: {
                               helperText: 'The MUI date picker component has migrated to @mui/x-date-picker.',
                             },
                             actionBar: {
-                              actions: ['today', 'clear']
-                            }
+                              actions: ['today', 'clear'],
+                            },
                           }}
                           label='Date picker'
                           value={selectedDate}
@@ -263,7 +265,7 @@ const Forms = () => {
                           slotProps={{
                             textField: {
                               helperText: 'The MUI time picker component has migrated to @mui/x-date-picker.',
-                            }
+                            },
                           }}
                           label='Time picker'
                           value={selectedDate}
@@ -274,7 +276,7 @@ const Forms = () => {
                       <Grid item xs={2}>
                         <KeyboardDatePicker
                           selectsRange
-                          textFieldProps={{label: 'Keyboard Range date picker', helperText: 'The range picker can be used by passing the `selectsRange` prop'}}
+                          textFieldProps={{ label: 'Keyboard Range date picker', helperText: 'The range picker can be used by passing the `selectsRange` prop' }}
                           selected={dateRange[0]}
                           startDate={dateRange[0]}
 
@@ -286,7 +288,7 @@ const Forms = () => {
                       </Grid>
                       <Grid item xs={2}>
                         <KeyboardDatePicker
-                          textFieldProps={{label: 'Keyboard date picker', helperText: 'This Datepicker supports keyboard inputs'}}
+                          textFieldProps={{ label: 'Keyboard date picker', helperText: 'This Datepicker supports keyboard inputs' }}
                           selected={selectedDate}
                           onChange={(date) => handleDateChange(date as Date)}
                           shouldOpenOnFocus
@@ -321,7 +323,7 @@ const Forms = () => {
                           onChange={event => setTextFieldValue(event === null ? '' : event.target.value)}
                           error={showError}
                           helperText='The icon can be implemented on start or end or both and should reflect error color'
-                          InputProps={{startAdornment: <InputAdornment position='start'><WarningIcon color={showError ? 'error' : 'inherit'}/></InputAdornment>}}
+                          InputProps={{ startAdornment: <InputAdornment position='start'><WarningIcon color={showError ? 'error' : 'inherit'} /></InputAdornment> }}
                         />
                       </Grid>
                       <Grid item xs={2}>
@@ -336,6 +338,100 @@ const Forms = () => {
                       </Grid>
                     </Grid>
                   </Box>
+                </Grid>
+                <Grid item container xs={12}>
+                  <Grid item xs={2}>
+                    <PriceField
+                      textFieldProps={{ id: `${textFieldENId}WithoutDecimals`, label: 'Price field EN', helperText: 'without decimals', required: true }}
+                      showDecimals={false}
+                      language={LANGUAGE.ENGLISH}
+                      currency='gbp'
+                      valueInCent={priceValue}
+                      setValue={value => {
+                        console.log(`setting showcase price for ${textFieldENId}WithDecimals`, value);
+                        setPriceValue(value)
+                      }}
+                    />
+                  </Grid>
+                  <Grid item xs={2}>
+                    <PriceField
+                      textFieldProps={{ id: `${textFieldDEId}WithoutDecimals`, label: 'Price field DE', helperText: 'without decimals', required: true }}
+                      showDecimals={false}
+                      selectOnFocus
+                      disabled
+                      language={LANGUAGE.GERMAN}
+                      valueInCent={priceValue}
+                      setValue={value => {
+                        console.log(`setting showcase price for ${textFieldENId}WithDecimals`, value);
+                        setPriceValue(value)
+                      }}
+                    />
+
+                  </Grid>
+                  <Grid item xs={2}>
+                    <PriceField
+                      textFieldProps={{ id: `${textFieldENId}WithDecimals`, label: 'Price field EN', helperText: 'with decimals', required: true }}
+                      showDecimals={true}
+                      language={LANGUAGE.ENGLISH}
+                      currency='gbp'
+                      currencyPosition={'start'}
+                      valueInCent={priceValue}
+                      setValue={value => {
+                        console.log(`setting showcase price for ${textFieldENId}WithDecimals`, value);
+                        setPriceValue(value)
+                      }}
+                    />
+                  </Grid>
+                  <Grid item xs={2}>
+                    <PriceField
+                      textFieldProps={{ id: `${textFieldDEId}WithDecimals`, label: 'Price field DE', helperText: 'with decimals', required: true }}
+                      showDecimals={true}
+                      selectOnFocus
+                      language={LANGUAGE.GERMAN}
+                      valueInCent={priceValue}
+                      currencyPosition={'start'}
+                      setValue={value => {
+                        console.log(`setting showcase price for ${textFieldENId}WithDecimals`, value);
+                        setPriceValue(value)
+                      }}
+                    />
+                  </Grid>
+                  <Grid item xs={2}>
+                    <PriceField
+                      textFieldProps={{ id: `${textFieldENId}WithNegativeMinMax`, label: 'Price field EN', helperText: 'with negative min/max', required: true }}
+                      showDecimals={true}
+                      min={-100000}
+                      max={-1000}
+                      language={LANGUAGE.ENGLISH}
+                      currency='gbp'
+                      valueInCent={priceValue}
+                      setValue={value => {
+                        console.log(`setting showcase price for ${textFieldENId}WithDecimals`, value);
+                        setPriceValue(value)
+                      }}
+                    />
+                  </Grid>
+                  <Grid item xs={2}>
+                    <PriceField
+                      textFieldProps={{ id: `${textFieldDEId}WithPositiveMinMax`, label: 'Price field DE', helperText: 'with positive min/max', required: true }}
+                      showDecimals={true}
+                      min={1000}
+                      max={100000}
+                      selectOnFocus
+                      language={LANGUAGE.GERMAN}
+                      valueInCent={priceValue}
+                      setValue={value => {
+                        console.log(`setting showcase price for ${textFieldENId}WithDecimals`, value);
+                        setPriceValue(value)
+                      }}
+                    />
+
+                  </Grid>
+                  <Grid item xs={2}>
+                    <Button sx={{ mt: 2 }} variant={'contained'} type={'submit'} onClick={(e) => {
+                      console.log('submitted button', { e, priceValueEnWithDecimals: priceValue })
+                    }}>submit price field test</Button>
+                  </Grid>
                 </Grid>
               </Grid>
             </Box>