+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [11.2.0] - 2023-12-08
+- Reworked `PriceInput` Component.
+  - PriceInput now emits a value of type `number`
+  - PriceInput now validates the input value and only emits valid values. 
+
+## [11.1.0]
+- replaced deprecated dependencies
+
 ## [11.0.0] - 2023-02-01
 - bump/add/remove core libraries and dependencies
 - reintegrated showcase into project to faciliate development